first_index() now returns size_t and terminates early if period == step

+++ libnestutil/numerics.h
@@ -177,6 +177,6 @@ long mod_inverse( long a, long m );
  *
  * See comments in the function definition for implementation details.
  */
-long first_index( long period, long phase0, long step, long phase );
+size_t first_index( long period, long phase0, long step, long phase );
 
 #endif
